backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
starlette
openai


numpy

scikit-learn
